lambda-functions/plain-lambda/requirements.txt
requests==2.31.0
python-dateutil==2.8.2
PyJWT==2.8.0 